Create the vendor dir and pass the path to bundler
The vendor directory needs to exist in the release_path so the symlink
can be created.

The path vendor/bundle should be passed into the bundle command so it
gets installed into the directory under the symlink and available across
deployment. Bundler will do the right thing if gems need to be updated.