Adds a configurable wait time to runit scripts
Because unicorn convention is to sleep 1 between forking, and runit has
a 7 second timeout for service start, running >7 unicorn children is
likely to cause a runit timeout.

This adds a new attribute `runit_wait_time` to the `application_ruby`
provider that allows setting the time runit will wait before timeing
out. It defaults to 7, which is the runit default.