Test on ubuntu 12.04 and centos 6.5 too.
This switches to runit because 1) bundle exec breaks poise-service's sysvinit implementation since when it execs you can't find it to get a pid anymore and 2) webrick under ruby 1.8 and 1.9 doesn't respond to SIGTERM so we need SIGKILL instead.

Doesn't add centos 7 because poise-service doesn't have systemd support yet so runit can't be started.