Rename BundlerMixin to RubyMixin since it does slightly more now.
All resources get a parent_ruby link and if the app doesn't have a 
bundle_install resource (not sure I'm still in love with the app_state
thing but leaving it for now) it will try to use the Ruby's executable
directory for the command. Seems better than nothing for now at least.

Thanks to @charlesjohnson for his support on Kickstarter!